requirements.txt
cryptography == 2.8
coreapi == 2.3.3
docutils == 0.13.1
django == 2.2.2
django-filter == 2.1.0
django-rest-swagger == 2.1.2
django_redis == 4.10.0
django-oauth-toolkit == 1.2.0
django-storages == 1.7.1
django-celery == 3.2.2
django-crispy_forms == 1.7.2
django-formtools == 2.1
django-import-export == 1.2.0
drf-writable-nested == 0.5.1
djangorestframework == 3.9.3
djangorestframework-jwt == 1.11.0
future == 0.17.1
inapppy == 2.3
markdown == 3.0.1
mysqlclient == 1.3.14
pyjwt == 1.7.1
pyyaml == 3.13
pygments == 2.3.1
redis == 2.10.6
requests_oauthlib == 1.2.0
six == 1.12.0
uwsgi == 2.0.18